Changes to Logo and CSS
Fixed an issue where I had accidently reverted the CSS back to old code causing the links not to sit correctly inside of the grid boxes on the portfolio page.

I also added a <a> tag to the logo that redirects back to the home page.

Further I used W3C validator to validate all of the code in the CSS and HTML files. I went through and made corrections as suggested.